+++ actix-multipart/src/error.rs
@@ -48,7 +48,9 @@ pub enum Error {
     /// always include a "name" parameter.
     ///
     /// [RFC 7578 §4.2]: https://datatracker.ietf.org/doc/html/rfc7578#section-4.2
-    #[display("Content-Disposition header was not found when parsing a \"form-data\" field")]
+    #[display(
+        "Content-Disposition 'name' parameter was not found when parsing a \"form-data\" field"
+    )]
     ContentDispositionNameMissing,
 
     /// Nested multipart is not supported.
@@ -72,7 +74,7 @@ pub enum Error {
     NotConsumed,
 
     /// Form field handler raised error.
-    #[display("An error occurred processing field: {name}")]
+    #[display("An error occurred processing field '{name}': {source}")]
     Field {
         name: String,
         source: actix_web::Error,
@@ -121,6 +123,21 @@ mod tests {
         assert_eq!(
             err.to_string(),
             "Error during field parsing: message is incomplete"
+        );
+
+        let err = Error::ContentDispositionNameMissing;
+        assert_eq!(
+            err.to_string(),
+            "Content-Disposition 'name' parameter was not found when parsing a \"form-data\" field"
+        );
+
+        let err = Error::Field {
+            name: "avatar".to_owned(),
+            source: actix_web::error::ErrorBadRequest("invalid file format"),
+        };
+        assert_eq!(
+            err.to_string(),
+            "An error occurred processing field 'avatar': invalid file format"
         );
     }
 }